microsservices/balances/internal/entity/client.go
package entity

import (
	"errors"
	"time"

	"github.com/google/uuid"
)

type Client struct {
	ID        string
	Name      string
	Accounts  []*Account
	CreatedAt time.Time
	UpdatedAt time.Time
}

func NewClient(name string) (*Client, error) {
	client := &Client{
		ID:        uuid.New().String(),
		Name:      name,
		CreatedAt: time.Now(),
		UpdatedAt: time.Now(),
	}
	err := client.Validate()

	if err != nil {
		return nil, err
	}

	return client, nil
}

func (c *Client) Validate() error {
	if c.Name == "" {
		return errors.New("name is required")
	}

	return nil
}

func (c *Client) Update(name string) error {
	c.Name = name
	c.UpdatedAt = time.Now()

	err := c.Validate()
	if err != nil {
		return err
	}

	return nil
}

func (c *Client) AddAccount(account *Account) error {
	if account.Client.ID != c.ID {
		return errors.New("account does not belong to this client")
	}

	c.Accounts = append(c.Accounts, account)
	return nil
}